+++ client/web/src/extensions/ExtensionsArea.tsx
@@ -78,7 +78,7 @@ export const ExtensionsArea: React.FunctionComponent<ExtensionsAreaProps> = prop
     }
 
     return (
-        <div className="extensions-area web-content mt-3">
+        <div className="extensions-area web-content mt-4">
             <ExtensionsAreaHeader
                 {...props}
                 {...context}