+++ migrations/frontend/1528395780_drop_secrets_sql.up.sql
@@ -1,0 +1,6 @@
+BEGIN;
+
+-- The `secrets` table is not used and has no records.
+DROP TABLE IF EXISTS secrets;
+
+COMMIT;